pyproject.toml
[tool.poetry]
name = "rutube-hackathon"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
beautifulsoup4 = "^4.12.3"
html2text = "^2024.2.26"
pandas = "^2.2.2"
langchain = "^0.3.1"
python-dotenv = "^1.0.1"
langchain-community = "^0.3.1"
openai = "^1.48.0"
deepeval = "^1.3.1"
mistralai = "^1.1.0"
# weaviate-client = "^4.8.1"
flashrank = "^0.2.9"
langchain-core = "^0.3.6"
ipywidgets = "^8.1.5"
# langchain-chroma = "^0.1.4"
# sentence-transformers = ">=2.2.2,<2.3.0"
sentence-transformers = ">=3.1.1"
# langchain-mistralai = "^0.2.0"

torch = [
    {platform = "darwin", version = "^2", source = "pypi"},
    {platform = "linux", version = "^2", source = "pypi"},
    {platform = "win32", version = "^2", source = "torch124"}
]
chromadb = [
    {version = "0.5.0", platform = "win32"},
    {version = "^0.5.0", platform = "linux"},
    {version = "^0.5.0", platform = "darwin"}
]
chroma-hnswlib = [
    {version = '0.7.3', platform = "win32"},
    {version = '^0.7.3', platform = "linux"},
    {version = '^0.7.3', platform = "darwin"}
]

openpyxl = "^3.1.5"
rapidfuzz = "^3.10.0"
optimum = {extras = ["onnxruntime-gpu"], version = "^1.22.0"}
onnxruntime-gpu = "1.16.1"

[[tool.poetry.source]]
name = "torch124"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
